Change IDs in database from ints to longs

+++ app/src/main/java/com/krystian/chessclock/model/CustomGame.kt
@@ -1,10 +1,11 @@
 package com.krystian.chessclock.model
 
 data class CustomGame(
-    val id: Int = 0,
+    val id: Long = 0,
     val whiteTime: Int,
     val whiteIncrement: Int,
     val blackTime: Int,
     val blackIncrement: Int,
-    val matchId: Int
+    val matchId: Long
 )
+
